requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.1.3
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.0rc1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==42.0.8
cssselect==1.3.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jieba3k==0.35.1
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
netmiko==4.4.0
newspaper3k==0.2.8
nltk==3.9.1
ntc_templates==6.0.0
packaging==25.0
paramiko==3.4.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
scp==0.15.0
setuptools==71.0.3
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.7
sqlparse==0.5.3
textfsm==1.1.3
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
